requirements.txt
absl-py==0.7.1
aiohttp==3.5.4
appdirs==1.4.3
arandr==0.1.10
asn1crypto==0.24.0
astor==0.8.0
async-timeout==3.0.1
attrs==19.1.0
btrfsutil==1.1.1
CacheControl==0.12.5
cachetools==3.1.1
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
cryptography==2.6.1
cupshelpers==1.0
cycler==0.10.0
decorator==4.4.0
distlib==0.2.9
distro==1.4.0
docopt==0.6.2
entrypoints==0.3
ffmpeg==1.4
gast==0.2.2
google-api-core==1.12.0
google-auth==1.6.3
google-cloud==0.34.0
google-cloud-language==1.2.0
google-cloud-vision==0.37.0
google-pasta==0.1.7
googleapis-common-protos==1.6.0
grpcio==1.21.1
gufw==18.10.0
h5py==2.9.0
html5lib==1.0.1
idna==2.8
jeepney==0.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==19.0.1
kiwisolver==1.1.0
lensfun==0.3.2
lockfile==0.12.2
louis==3.10.0
Markdown==3.1.1
matplotlib==3.1.0
msgpack==0.6.1
multidict==4.5.2
npyscreen==4.10.5
numpy==1.16.4
opencv-python==4.1.0.25
packaging==19.0
pep517==0.5.0
Pillow-SIMD==5.3.0.post1
ply==3.11
progress==1.5
protobuf==3.8.0
psutil==5.6.2
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycairo==1.18.1
pycparser==2.19
pycups==1.9.74
pycurl==7.43.0.2
PyGObject==3.32.1
pyparsing==2.4.0
PyQt5==5.12.2
PyQt5-sip==4.19.17
pyte==0.8.0
python-dateutil==2.8.0
python-libtorrent==1.1.12
python-xapp==1.6.0
python-xlib==0.24
pytoml==0.1.20
pytz==2019.1
ranger-fm==1.9.2
requests==2.22.0
retrying==1.3.3
rsa==4.0
SecretStorage==3.1.1
setproctitle==1.1.10
six==1.12.0
speedtest-cli==2.1.1
syslogng==1.0
team==1.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
thefuck==3.29
ueberzug==18.1.3
ufw==0.36
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
websockets==6.0
Werkzeug==0.15.4
wrapt==1.11.2
yarl==1.3.0
